+++ NEWS.md
@@ -1,3 +1,9 @@
+# ggview 0.2.0
+* Implemented `canvas()` element to be used instead of running `ggview()` 
+directly.
+* Added zoom and pan functionality.
+* Added `save_ggplot()` function to save a ggplot with canvas parameters.
+
 # ggview 0.1.0
 
 * Initialize package.